test fix

+++ wsb-jpa1/src/main/java/com/jpacourse/persistence/dao/PatientDao.java
@@ -1,13 +1,6 @@
 package com.jpacourse.persistence.dao;
 
-import com.jpacourse.persistence.entity.AddressEntity;
 import com.jpacourse.persistence.entity.PatientEntity;
-import com.jpacourse.persistence.entity.VisitEntity;
-import com.jpacourse.persistence.entity.DoctorEntity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
@@ -17,6 +10,6 @@ import java.util.List;
 public interface PatientDao extends Dao<PatientEntity, Long> {
     void addVisit(Long patientId, Long doctorId, String visitDescription, LocalDateTime visitDate, Long treatmentId);
     List<PatientEntity> findByLastName(String lastName);
-    List<PatientEntity> findByVisitsGreaterThan(int visitCount);
+    List<PatientEntity> findByVisitsGreaterThan(long visitCount);
     List<PatientEntity> findByAgeGreaterThan(Integer age);
 }
